Moves platform check to root
