+++ src/searchBarPlatforms.tsx
